🐛 속성값 삭제 시 convertAttribute 적용
