+++ src/library/dom/attribute.js
@@ -1,5 +1,24 @@
-export const isEventAttribute = (attribute, value) =>
-  attribute.startsWith('on') && typeof value === 'function';
+export const updateAttributes = (target, newProps, oldProps) => {
+  Object.entries(oldProps).forEach(([attr, value]) => {
+    if (isEventAttribute(attr, value)) {
+      const event = attr.slice(2).toLowerCase();
+      return target.removeEventListener(event, value);
+    }
+    if (newProps[attr] === undefined) {
+      target.removeAttribute(convertAttribute(attr));
+    }
+  });
+
+  Object.entries(newProps).forEach(([attr, value]) => {
+    if (oldProps[attr] === newProps[attr]) {
+      return;
+    }
+    setAttribute(target, attr, value);
+  });
+};
+
+const isEventAttribute = (attribute, value) =>
+  attribute.startsWith("on") && typeof value === "function";
 
 export const setAttribute = (element, attr, value) => {
   if (isEventAttribute(attr, value)) {
@@ -10,11 +29,12 @@ export const setAttribute = (element, attr, value) => {
 };
 
 const convertAttribute = (attribute) => {
-  if (attribute === 'className') {
-    return 'class';
+  switch (attribute) {
+    case "className":
+      return "class";
+    case "htmlFor":
+      return "for";
+    default:
+      return attribute;
   }
-  if (attribute === 'htmlFor') {
-    return 'for';
-  }
-  return attribute;
 };
